backend/requirements.txt
flask
flask-cors
psycopg2-binary
flasgger
